moved cell splitting into a separate class
